ページ遷移時の値渡しをContext使用に変更
